[WIP] To support batch concurrent download of segments
The idea is to collect all segment URLs at an earlier stage,
use gevent to create a number of threads (defaults to 6), fetch them
all using requests with TCP session reuse.

Because this feature is radically different from the existing framework,
it is difficult to merge it "naturally". We shall improve it gradually.